% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{calc_Fj}
\alias{calc_Fj}
\title{Calcula o estimador Fj}
\usage{
calc_Fj(C, j, I = 52)
}
\arguments{
\item{C}{Matriz cumulativa}

\item{j}{Semana de atraso}

\item{I}{Total de semanas}
}
\value{
Valor de Fj
}
\description{
Calcula o estimador Fj
}
