% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{calc_lj}
\alias{calc_lj}
\title{Calcula a largura fuzzy lj associada ao estimador Fj}
\usage{
calc_lj(C, X, j)
}
\arguments{
\item{C}{Matriz cumulativa.}

\item{X}{Matriz de atraso (frequências).}

\item{j}{Índice do atraso.}
}
\value{
Valor escalar lj.
}
\description{
Calcula a largura fuzzy lj associada ao estimador Fj
}
