% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{calcular_atraso}
\alias{calcular_atraso}
\title{Calcula o atraso de notificação em semanas}
\usage{
calcular_atraso(dados)
}
\arguments{
\item{dados}{Base de dados com datas}
}
\value{
Base com coluna de atraso
}
\description{
Calcula o atraso de notificação em semanas
}
