% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{estimacao_delay_fuzzy}
\alias{estimacao_delay_fuzzy}
\title{Correção de atraso por estimadores fuzzy}
\usage{
estimacao_delay_fuzzy(dados, data_notificacao, data_sintomas)
}
\arguments{
\item{dados}{Base de dados.}

\item{data_notificacao}{Coluna de notificação.}

\item{data_sintomas}{Coluna de início dos sintomas.}
}
\value{
Lista com matrizes intermediárias, estimadores e previsões fuzzy.
}
\description{
Correção de atraso por estimadores fuzzy
}
