% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{estimar_Fj}
\alias{estimar_Fj}
\title{Estima o fator de correção Fj}
\usage{
estimar_Fj(C, j)
}
\arguments{
\item{C}{Matriz cumulativa de atrasos.}

\item{j}{Índice do atraso (semana).}
}
\value{
Valor escalar do estimador Fj.
}
\description{
Estima o fator de correção Fj
}
