% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{fuzzy_Fj}
\alias{fuzzy_Fj}
\title{Gera números fuzzy triangulares}
\usage{
fuzzy_Fj(Fjs, ljs)
}
\arguments{
\item{Fjs}{Vetor de Fj}

\item{ljs}{Vetor de larguras}
}
\value{
Lista de números fuzzy
}
\description{
Gera números fuzzy triangulares
}
