% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{matriz_atraso}
\alias{matriz_atraso}
\title{Cria matriz de atrasos por semana epidemiológica}
\usage{
matriz_atraso(dados)
}
\arguments{
\item{dados}{Base com atraso calculado}
}
\value{
Matriz de atraso
}
\description{
Cria matriz de atrasos por semana epidemiológica
}
