% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{matriz_cumulativa}
\alias{matriz_cumulativa}
\title{Gera matriz cumulativa de atrasos}
\usage{
matriz_cumulativa(matriz)
}
\arguments{
\item{matriz}{Matriz de atraso}
}
\value{
Matriz cumulativa
}
\description{
Gera matriz cumulativa de atrasos
}
