% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{mult_fuzzy}
\alias{mult_fuzzy}
\title{Multiplicação de números fuzzy triangulares}
\usage{
mult_fuzzy(A, B)
}
\arguments{
\item{A}{Vetor (centro, esquerda, direita).}

\item{B}{Vetor (centro, esquerda, direita).}
}
\value{
Vetor fuzzy triangular resultante.
}
\description{
Multiplicação de números fuzzy triangulares
}
