% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{propagar_fuzzy}
\alias{propagar_fuzzy}
\title{Propaga incerteza fuzzy ao longo da matriz cumulativa}
\usage{
propagar_fuzzy(C, Fjs, ljs)
}
\arguments{
\item{C}{Matriz cumulativa.}

\item{Fjs}{Vetor de fatores Fj.}

\item{ljs}{Vetor de larguras fuzzy lj.}
}
\value{
Lista com matrizes centro, esquerda e direita.
}
\description{
Propaga incerteza fuzzy ao longo da matriz cumulativa
}
