% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq_funcoes.R
\name{calcular_atraso}
\alias{calcular_atraso}
\title{Calcula o atraso de notificação em semanas}
\usage{
calcular_atraso(dados, data_notificacao, data_sintomas)
}
\arguments{
\item{dados}{Base de dados}

\item{data_notificacao}{Coluna da data de notificação}

\item{data_sintomas}{Coluna da data de início dos sintomas}
}
\value{
Base com coluna atrasoSemanas
}
\description{
Calcula o atraso de notificação em semanas
}
